move sessionbook settings
