+++ app/src/features/settings/components/SettingsIndex/index.scss
@@ -2,7 +2,7 @@
   width: 100vw;
   height: 100%;
   display: flex;
-  gap: 24px;
+  gap: 20px;
   background: var(--surface-0);
 }
 
@@ -11,7 +11,7 @@
   justify-content: center;
   height: 100%;
   flex: 1;
-  padding: 1rem 0;
+  padding: 1rem;
 
   & .settings-content {
     width: 100%;
